requirements.txt
flask==3.1.2
flask-cors==6.0.1
tensorflow==2.20.0
numpy==1.26.4
pillow
